+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [PEP-440](https://www.python.org/dev/peps/pep-0440/)
 
 ### Added
 
+- Send Auth notifications for user requests to get blueprints copies (#7)
+
 ### Changed
 
 - Added CI tests for AA 2.9 / Django 3.2